Implement dummy api

+++ backend-go/api/codepair/v1/models/model_http_exception_response.go
@@ -3,7 +3,7 @@ package models
 type HttpExceptionResponse struct {
 
 	// HTTP status code
-	StatusCode float32 `json:"statusCode"`
+	StatusCode int32 `json:"statusCode"`
 
 	// Description of the error
 	Message string `json:"message"`
